Cleaned up fb methods
