+++ app/src/main/res/layout/activity_main.xml
@@ -9,17 +9,4 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="adi.sf1.targaryen.newyorktimes.MainActivity">
 
-
-    <com.facebook.share.widget.ShareButton
-        android:id="@+id/fb_share_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/fb_login_button"
-        android:layout_margin="10dp"
-        />
-    <com.facebook.login.widget.LoginButton
-        android:id="@+id/fb_login_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"/>
-
 </RelativeLayout>
